remove and edit comments

+++ app/server/cardOperations.ts
@@ -118,12 +118,22 @@ export async function getAttachments({ cardId }: { cardId: string }) {
 
 export async function removeAttachment({
 	cardId,
-	attachmentId
+	attachmentId,
+	attachmentAuthor,
+	cardAuthor,
+	userId
 }: {
 	cardId: string
 	attachmentId: string
+	attachmentAuthor: string
+	cardAuthor: string
+	userId: string
 }) {
 	try {
+		if (userId !== attachmentAuthor || userId !== cardAuthor) {
+			throw new Error('Unauthorized')
+		}
+
 		const removeAttachment = await prisma.attachment.delete({
 			where: {
 				cardId,
@@ -136,5 +146,109 @@ export async function removeAttachment({
 	} catch (e) {
 		console.error(e)
 		return { e }
+	}
+}
+
+export async function getComments({ cardId }: { cardId: string }) {
+	try {
+		const comments = await prisma.comment.findMany({
+			where: {
+				cardId
+			}
+		})
+
+		return { comments }
+	} catch (e) {
+		console.error(e)
+		return { e }
+	}
+}
+
+export async function addComment({
+	cardId,
+	authorId,
+	comment
+}: {
+	cardId: string
+	authorId: string
+	comment: string
+}) {
+	try {
+		const addComment = await prisma.comment.create({
+			data: {
+				cardId,
+				authorId,
+				text: comment
+			}
+		})
+
+		return { addComment }
+	} catch (e) {
+		console.error(e)
+		return { e }
+	}
+}
+
+export async function removeComment({
+	cardId,
+	commentId,
+	commentAuthor,
+	cardAuthor,
+	userId
+}: {
+	cardId: string
+	commentId: string
+	commentAuthor: string
+	cardAuthor: string
+	userId: string
+}) {
+	try {
+		if (userId !== commentAuthor || userId !== cardAuthor) {
+			throw new Error('Unauthorized')
+		}
+
+		await prisma.comment.delete({
+			where: {
+				id: commentId,
+				cardId
+			}
+		})
+	} catch (e) {
+		console.error(e)
+		return (e as Error).message
+	}
+}
+
+export async function updateComment({
+	cardId,
+	commentId,
+	authorId,
+	userId,
+	comment
+}: {
+	cardId: string
+	commentId: string
+	authorId: string
+	userId: string
+	comment: string
+}) {
+	try {
+		if (userId !== authorId) {
+			throw new Error('Unauthorized')
+		}
+
+		const updatedComment = await prisma.comment.update({
+			where: {
+				cardId,
+				id: commentId
+			},
+			data: {
+				text: comment
+			}
+		})
+		return { updatedComment }
+	} catch (e) {
+		console.error((e as Error).message)
+		return (e as Error).message
 	}
 }
